Create BasicLights.aspx
This is the basic light code. No CSS styling. Page shows a picture of the object with CheerLights and then visual representations of both the current CheerLights color and the one pulled from our feed. Please see the BasicLights.aspx.cs file for the code to update the light color.